Delete unused elevation.py
